+++ gestion-tienda/src/main/java/com/tutienda/gestion_tienda/service/VentaService.java
@@ -1,5 +1,6 @@
 package com.tutienda.gestion_tienda.service;
 
+import com.tutienda.gestion_tienda.exception.ResourceNotFoundException;
 import com.tutienda.gestion_tienda.models.Venta;
 import com.tutienda.gestion_tienda.repository.VentaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,9 +20,10 @@ public class VentaService {
         return ventaRepository.findAll();
     }
 
-    // Obtener una venta por ID
-    public Optional<Venta> obtenerVentaPorId(Long id) {
-        return ventaRepository.findById(id);
+    // Obtener una venta por ID con manejo de excepción
+    public Venta obtenerVentaPorId(Long id) {
+        return ventaRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Venta con ID " + id + " no encontrado"));
     }
 
     // Guardar una nueva venta
@@ -29,8 +31,11 @@ public class VentaService {
         return ventaRepository.save(venta);
     }
 
-    // Eliminar una venta
+    // Eliminar una venta con manejo de excepción
     public void eliminarVenta(Long id) {
+        if (!ventaRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Venta con ID " + id + " no encontrado");
+        }
         ventaRepository.deleteById(id);
     }
 }
